requirements.txt
langchain-google-genai
langgraph
pydantic
python-dotenv
duckduckgo-search
langchain-community